requirements.txt
python-telegram-bot
mysql-connector-python
confuse